feat: move LevelSelection to its own file
